requirements.txt
fastapi
uvicorn
tortoise-orm
aiopg
paho-mqtt
python-dotenv
asyncpg
asyncio